flaskapp/requirements.txt
Flask==3.1.0
PyYAML==6.0.2
mysql-connector-python==9.1.0
jsonrpc
sqlalchemy
